+++ familyhelper-webapi-stack/src/main/java/com/dwarfeng/familyhelper/webapi/stack/service/notify/SendExecutorSupportResponseService.java
@@ -1,6 +1,6 @@
 package com.dwarfeng.familyhelper.webapi.stack.service.notify;
 
-import com.dwarfeng.notify.stack.bean.entity.SenderSupport;
+import com.dwarfeng.familyhelper.plugin.notify.bean.entity.SendExecutorSupport;
 import com.dwarfeng.subgrade.stack.bean.dto.PagedData;
 import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
@@ -8,18 +8,16 @@ import com.dwarfeng.subgrade.stack.exception.ServiceException;
 import com.dwarfeng.subgrade.stack.service.Service;
 
 /**
- * 发送器支持响应服务。
+ * 发送执行器支持响应服务。
  *
  * @author DwArFeng
- * @since 1.0.6
+ * @since 1.0.7
  */
-public interface SenderSupportResponseService extends Service {
+public interface SendExecutorSupportResponseService extends Service {
 
     boolean exists(StringIdKey key) throws ServiceException;
 
-    SenderSupport get(StringIdKey key) throws ServiceException;
+    SendExecutorSupport get(StringIdKey key) throws ServiceException;
 
-    PagedData<SenderSupport> all(PagingInfo pagingInfo) throws ServiceException;
-
-    PagedData<SenderSupport> idLike(String pattern, PagingInfo pagingInfo) throws ServiceException;
+    PagedData<SendExecutorSupport> all(PagingInfo pagingInfo) throws ServiceException;
 }