+++ CHANGELOG.md
@@ -37,6 +37,7 @@
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.note.NoteItemController。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.note.NoteNodeController。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.note.PonbController。
+  - com.dwarfeng.familyhelper.webapi.node.controller.v1.notify.SendExecutorSupportController。
 
 #### Bug修复
 
@@ -44,7 +45,12 @@
 
 #### 功能移除
 
-- (无)
+- 删除 Controller。
+  - com.dwarfeng.familyhelper.webapi.node.controller.v1.notify.RouterInfoController。
+  - com.dwarfeng.familyhelper.webapi.node.controller.v1.notify.RouterSupportController。
+  - com.dwarfeng.familyhelper.webapi.node.controller.v1.notify.SenderInfoController。
+  - com.dwarfeng.familyhelper.webapi.node.controller.v1.notify.SenderRelationController。
+  - com.dwarfeng.familyhelper.webapi.node.controller.v1.notify.SenderSupportController。
 
 ---